+++ crates/modular_core/src/dsp/filters/lowpass.rs
@@ -113,7 +113,7 @@ impl LowpassFilter {
     }
 
     fn update(&mut self, sample_rate: f32) {
-        let channels = self.channel_count() as u8;
+        let channels = self.channel_count();
 
         self.outputs.sample.set_channels(channels);
