+++ crates/modular_core/src/dsp/oscillators/d_pulse.rs
@@ -52,7 +52,7 @@ impl DPulseOscillator {
     fn update(&mut self, _sample_rate: f32) {
         let num_channels = self.channel_count();
 
-        self.outputs.sample.set_channels(num_channels as u8);
+        self.outputs.sample.set_channels(num_channels);
 
         for ch in 0..num_channels {
             let state = &mut self.channels[ch];
