+++ crates/modular_core/src/dsp/utilities/remap.rs
@@ -43,7 +43,7 @@ pub struct Remap {
 
 impl Remap {
     fn update(&mut self, _sample_rate: f32) {
-        let channels = self.channel_count() as u8;
+        let channels = self.channel_count();
 
         self.outputs.sample.set_channels(channels);
